pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "queso"
version = "0.0.0"
requires-python = ">=3.9"
dependencies = [
    "h5py",  # data format for saving
    "matplotlib",
    "polars",  # post-analysis dataframes
    "seaborn",
    "numpy",
    "jax",  # differentiable programming
    "flax",  # neural network library
    "tensorcircuit",  # quantum circuits
    "qiskit",
    "pylatexenc",
#    "torch",
    "optax",  # optimization management
    "python-dotenv",  # local environment variables
    "prettytable",
    "psutil",
    "GPUtil",
]


a = ["mkdocs", "mkdocstrings-python", "mkdocs-markdownextradata-plugin", "mkdocs-jupyter"]